Update 2015-09-25-network-analysis.md
Overall, I thought this was clear and easy to follow.  I added a few more links for terms people may not be familiar with.  I also made some copy edits.  

One idea for line 25 would be to reference the workshop done on Github in September.  I'm not sure how to link the two pages together, but some sort of hyperlink or additional reference to the other workshop might be quite useful to the reader.