Update 2015-11-20-visualizing-data.md
I really enjoyed reading this workshop!  I learned a lot! 
I just made a few changes so it would be accessible to an audience that was not at the workshop and just seeing the information online.  Please feel free to keep or discard these changes as you see fit. 

Thanks,
Laura