Update 2016-01-15-web-annotations.md
I really learned a lot from this post!  It's very helpful.
Starting on line 36, I started to get very confused.  I wasn't really clear between the difference between Annotations (line 41) and An annotation (line 43).  Are these the two definitions referenced in line 38?  As someone who is not familiar with this set of online tools, I felt very lost.  A little more explanation or clarity would be very useful.  
Fixed a few typos and tried to make things clearer for a reader who might not be familiar with these programs.
Thanks! Laura